requirements.txt
django
Jinja2
psycopg2-binary
python-decouple
djangorestframework
